impl-newcmd
Added abunch of cool stuff. Yeah.
