+++ Aura Operating System/Aura OS/Shell/cmdIntr/FileSystem/Chmod.cs
@@ -1,0 +1,9 @@
+﻿namespace Aura_OS.Shell.cmdIntr.FileSystem
+{
+    /// <summary>
+    /// I added this is, this is useless until you code it.
+    /// </summary>
+    class Chmod
+    {
+    }
+}
