+++ Aura Operating System/Aura OS/Shell/cmdIntr/SystemInfomation/Version.cs
@@ -1,0 +1,37 @@
+﻿/*
+* PROJECT:          Aura Operating System Development
+* CONTENT:          Command Interpreter - Version
+* PROGRAMMER(S):    John Welsh <djlw78@gmail.com>
+*/
+
+using L = Aura_OS.System.Translation;
+
+namespace Aura_OS.Shell.cmdIntr.SystemInfomation
+{
+    class Version
+    {
+        private static string HelpInfo = "";
+
+        /// <summary>
+        /// Getter and Setters for Help Info.
+        /// </summary>
+        public static string HI
+        {
+            get { return HelpInfo; }
+            set { HelpInfo = value; /*PUSHED OUT VALUE (in)*/}
+        }
+
+        /// <summary>
+        /// Empty constructor. (Good for debug)
+        /// </summary>
+        public Version() { }
+
+        /// <summary>
+        /// c = command, c_Version
+        /// </summary>
+        public static void c_Version()
+        {
+            L.Text.Display("about");
+        }
+    }
+}
